Update kubeadm_init.yaml
For single Control Plane setup, add --apiserver-advertise-address with the local/private IP address to init command. Otherwise, the join command could use an incorrect IP adress (NAT IP address will be used if it is in the defult route of the control plane server). The workers are not able to join the cluster by using the NAT IP address.